Утилиты вынесены в отдельную папку и разделены по ответственности
