+++ src/mock/event-item.js
@@ -1,5 +1,18 @@
-import {getRandomBool, getRandomElement, getRandomInteger, getRandomISODates, getRandomOffers} from './utils';
+import {
+  getRandomBool,
+  getRandomElement,
+  getRandomInteger,
+  getRandomISODates
+} from '../utils/mock';
 import {DESTINATIONS, TYPES} from '../const';
+
+const OFFERS = [
+  {title: 'Add luggage', price: 50},
+  {title: 'Switch to comfort', price: 80},
+  {title: 'Add breakfast', price: 50},
+  {title: 'Choose seats', price: 20},
+  {title: 'Travel by train', price: 120},
+];
 
 const LOREM = [
   'Lorem ipsum dolor sit amet, consectetur adipiscing elit.',
@@ -9,6 +22,10 @@ const LOREM = [
   'Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante.',
 ].join(' ');
 
+
+const getRandomOffers = (offersCount) => {
+  return new Array(offersCount).fill().map(() => getRandomElement(OFFERS));
+};
 
 const getRandomType = () => {
   return getRandomElement(TYPES);
@@ -48,4 +65,15 @@ const generateEventItem = () => {
   };
 };
 
-export {generateEventItem};
+const generateAvailableOffers = () => {
+  const offersCount = getRandomInteger(0, 3);
+
+  return TYPES.map((type) => {
+    return {
+      type: type,
+      offers: new Array(offersCount).fill().map(() => getRandomElement(OFFERS)),
+    };
+  });
+};
+
+export {generateEventItem, generateAvailableOffers};
